requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.30
python-dotenv==1.0.1
